# jara-market-backend/jaraman/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jaramarket API",
        "description": "Jaramarket Meals Marketplace API for Customer and Vendor mobile applications",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "L5_SWAGGER_CONST_HOST",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/vendors/{id}/reviews": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Get Vendor Reviews",
                "description": "Retrieve all reviews for a specific vendor.",
                "operationId": "a1149669556b3b73b22bf4900eccae6d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The Vendor ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reviews retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Vendor not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Submit Review for Vendor",
                "description": "Submit a rating and comment for a specific vendor.",
                "operationId": "5a3cc3b2d6714291bbdf69149df8f325",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The Vendor ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rating"
                                ],
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "example": 5,
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Excellent service and delicious food!",
                                        "maxLength": 1000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review submitted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Vendor not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/vendor/customers/{id}/reviews": {
            "post": {
                "tags": [
                    "Vendor"
                ],
                "summary": "Review a Customer",
                "description": "Vendors can leave a review and rating for a customer after a transaction.",
                "operationId": "40cd0eb90799a7a7dc935b6e068ab6b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The Customer ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rating"
                                ],
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "example": 5,
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Very polite customer, quick payment."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review submitted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Customer not found"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/customers/{id}/reviews": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Get Customer Reviews",
                "description": "Retrieve all reviews received by a customer.",
                "operationId": "c2c0e30040cd1335e908b868e4710bde",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The Customer ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer reviews retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Customer not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/jaram/social/{provider}": {
            "post": {
                "tags": [
                    "Customer Authentication",
                    "Vendor Authentication"
                ],
                "summary": "Social Authentication",
                "description": "Authenticate using a social provider (google, facebook, apple). The mobile app should provide the token obtained from the provider.",
                "operationId": "bda693eba800c3c8132f1b59311f65e1",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "description": "The social provider",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "google",
                                "facebook",
                                "apple"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token"
                                ],
                                "properties": {
                                    "token": {
                                        "description": "The access token or ID token from the provider",
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "customer",
                                        "enum": [
                                            "customer",
                                            "vendor"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authentication successful"
                    },
                    "400": {
                        "description": "Invalid token or provider"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}